+++ packages/examples/src/index.ts
@@ -22,7 +22,15 @@ import quadrantChart from './examples/quadrant-chart.js';
 import packetDiagram from './examples/packet.js';
 import blockDiagram from './examples/block.js';
 import treemapDiagram from './examples/treemap.js';
+import eventmodelingDiagram from './examples/eventmodeling.js';
 import vennDiagram from './examples/venn.js';
+import treeViewDiagram from './examples/tree-view.js';
+import wardleyDiagram from './examples/wardley.js';
+import cynefinDiagram from './examples/cynefin.js';
+import railroadDiagram from './examples/railroad.js';
+import railroadEbnfDiagram from './examples/railroad-ebnf.js';
+import railroadAbnfDiagram from './examples/railroad-abnf.js';
+import railroadPegDiagram from './examples/railroad-peg.js';
 
 export const diagramData: DiagramMetadata[] = [
   flowChart,
@@ -48,5 +56,13 @@ export const diagramData: DiagramMetadata[] = [
   packetDiagram,
   blockDiagram,
   treemapDiagram,
+  eventmodelingDiagram,
   vennDiagram,
+  treeViewDiagram,
+  wardleyDiagram,
+  cynefinDiagram,
+  railroadDiagram,
+  railroadEbnfDiagram,
+  railroadAbnfDiagram,
+  railroadPegDiagram,
 ];