+++ packages/mermaid/src/diagrams/er/styles.ts
@@ -1,5 +1,5 @@
 import * as khroma from 'khroma';
-import type { FlowChartStyleOptions } from '../flowchart/styles.js';
+import type { DiagramStylesProvider } from '../../diagram-api/types.js';
 
 const fade = (color: string, opacity: number) => {
   // @ts-ignore TODO: incorrect types from khroma
@@ -12,9 +12,37 @@ const fade = (color: string, opacity: number) => {
   // @ts-ignore incorrect types from khroma
   return khroma.rgba(r, g, b, opacity);
 };
+const COLOR_THEMES = new Set(['redux-color', 'redux-dark-color']);
 
-const getStyles = (options: FlowChartStyleOptions) =>
-  `
+const genColor: DiagramStylesProvider = (options) => {
+  const { theme, look, bkgColorArray, borderColorArray } = options;
+  if (!COLOR_THEMES.has(theme)) {
+    return '';
+  }
+  const hasBkgColors = bkgColorArray?.length > 0;
+  let sections = '';
+
+  for (let i = 0; i < options.THEME_COLOR_LIMIT; i++) {
+    sections += `
+
+    [data-look="${look}"][data-color-id="color-${i}"].node path {
+    stroke: ${borderColorArray[i]};
+    ${hasBkgColors ? `fill: ${bkgColorArray[i]};` : ''}
+    }
+
+    [data-look="${look}"][data-color-id="color-${i}"].node  rect {
+    stroke: ${borderColorArray[i]};
+    ${hasBkgColors ? `fill: ${bkgColorArray[i]};` : ''}
+     }
+    `;
+  }
+  return sections;
+};
+
+const getStyles: DiagramStylesProvider = (options) => {
+  const { look, theme, erEdgeLabelBackground, strokeWidth } = options;
+  return `
+    ${genColor(options)}
   .entityBox {
     fill: ${options.mainBkg};
     stroke: ${options.nodeBorder};
@@ -30,7 +58,17 @@ const getStyles = (options: FlowChartStyleOptions) =>
   }
 
   .labelBkg {
-    background-color: ${fade(options.tertiaryColor, 0.5)};
+    background-color: ${COLOR_THEMES.has(theme) && erEdgeLabelBackground ? erEdgeLabelBackground : fade(options.tertiaryColor, 0.5)};
+  }
+
+  .edgeLabel {
+    background-color: ${COLOR_THEMES.has(theme) && erEdgeLabelBackground ? erEdgeLabelBackground : options.edgeLabelBackground};
+  }
+  .edgeLabel .label rect {
+    fill: ${COLOR_THEMES.has(theme) && erEdgeLabelBackground ? erEdgeLabelBackground : options.edgeLabelBackground};
+  }
+  .edgeLabel .label text {
+    fill: ${options.textColor};
   }
 
   .edgeLabel .label {
@@ -54,12 +92,12 @@ const getStyles = (options: FlowChartStyleOptions) =>
   {
     fill: ${options.mainBkg};
     stroke: ${options.nodeBorder};
-    stroke-width: 1px;
+    stroke-width: ${look === 'neo' ? strokeWidth : '1px'};
   }
 
   .relationshipLine {
     stroke: ${options.lineColor};
-    stroke-width: 1;
+    stroke-width: ${look === 'neo' ? strokeWidth : '1px'};
     fill: none;
   }
 
@@ -68,16 +106,10 @@ const getStyles = (options: FlowChartStyleOptions) =>
     stroke: ${options.lineColor} !important;
     stroke-width: 1;
   }
-    
-  .edgeLabel {
-    background-color: ${options.edgeLabelBackground};
-  }
-  .edgeLabel .label rect {
-    fill: ${options.edgeLabelBackground};
-  }
-  .edgeLabel .label text {
-    fill: ${options.textColor};
+  [data-look=neo].labelBkg {
+    background-color: ${fade(options.tertiaryColor, 0.5)};
   }
 `;
+};
 
 export default getStyles;