+++ packages/parser/src/language/railroad-abnf/tokenBuilder.ts
@@ -1,0 +1,7 @@
+import { AbstractMermaidTokenBuilder } from '../common/index.js';
+
+export class RailroadAbnfTokenBuilder extends AbstractMermaidTokenBuilder {
+  constructor() {
+    super(['railroad-abnf']);
+  }
+}